split media module
