+++ build.gradle
@@ -1,6 +1,8 @@
 // Top-level build file where you can add configuration options common to all sub-projects/modules.
 
-buildscript {
+buildscript { 
+    ext.kotlin_version = '1.3.41'
+
     ext {
         kotlin_version = '1.3.41'
         coroutine_version = '1.1.1'
@@ -41,7 +43,7 @@ buildscript {
         maven { url 'https://kotlin.bintray.com/kotlin-eap' }
     }
     dependencies {
-        classpath 'com.android.tools.build:gradle:3.5.0-beta05'
+        classpath 'com.android.tools.build:gradle:3.5.0-rc01'
         classpath "org.jetbrains.kotlin:kotlin-gradle-plugin:$kotlin_version"
         // NOTE: Do not place your application dependencies here; they belong
         // in the individual module build.gradle files
