+++ feature/media/src/main/res/drawable/ic_play_circle_outline.xml
