+++ feature/media/src/main/res/drawable/ld_play_icon.xml
