+++ feature/media/src/main/res/values/attr_media_container.xml
@@ -1,0 +1,29 @@
+<?xml version="1.0" encoding="utf-8"?><!--
+  ~ Copyright (c) 2019. Matsuda, Akihit (akihito104)
+  ~
+  ~ Licensed under the Apache License, Version 2.0 (the "License");
+  ~ you may not use this file except in compliance with the License.
+  ~ You may obtain a copy of the License at
+  ~
+  ~     http://www.apache.org/licenses/LICENSE-2.0
+  ~
+  ~ Unless required by applicable law or agreed to in writing, software
+  ~ distributed under the License is distributed on an "AS IS" BASIS,
+  ~ WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+  ~ See the License for the specific language governing permissions and
+  ~ limitations under the License.
+  -->
+
+<resources>
+    <attr name="gapWidth" format="dimension|reference" />
+
+    <declare-styleable name="MediaThumbnailContainer">
+        <attr name="gapWidth" />
+    </declare-styleable>
+
+    <declare-styleable name="MediaThumbnailContainerTheme">
+        <attr name="mediaThumbnailContainerStyle" format="reference" />
+    </declare-styleable>
+
+    <style name="Widget.MediaThumbnailContainer" parent="@android:style/Widget" />
+</resources>
