+++ feature/timeline/build.gradle
@@ -33,6 +33,7 @@ android {
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
     implementation project(':candysh')
+    implementation project(':feature:media')
     implementation project(':data:repository')
     implementation project(':data:db')
     implementation project(':data:restclient')
